+++ src/index.ts
@@ -1,13 +1,26 @@
-import byString from "./sortables/byString";
-import byNumber from "./sortables/byNumber";
 import byDate from "./sortables/byDate";
 import byValue from "./sortables/byValue";
+import byString from "./sortables/byString";
+import byNumber from "./sortables/byNumber";
+import byStringLowerCase from "./sortables/byStringLowerCase";
+import sortAsync, { AsyncArray } from "./sortables/byAsyncValue";
 
 export default {
+  byDate,
+  byValue,
   byString,
   byNumber,
-  byDate,
-  byValue
+  sortAsync,
+  AsyncArray,
+  byStringLowerCase
 };
 
-export { byString, byNumber, byDate, byValue };
+export {
+  byDate,
+  byValue,
+  byString,
+  byNumber,
+  sortAsync,
+  AsyncArray,
+  byStringLowerCase
+};