+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.2.12](https://github.com/cosimochellini/sort-es/compare/v0.2.11...v0.2.12) (2020-03-03)
+
+
+### Features
+
+* added async sort ([284831c](https://github.com/cosimochellini/sort-es/commit/284831cf9333d7ce1c14a31c76c905bf1a00d766))
+* added byAny and byAsyncValue ([04dadcb](https://github.com/cosimochellini/sort-es/commit/04dadcb192af18795d1aa4de22961e350461eb3b))
+* added more sort options ([bdcdd9f](https://github.com/cosimochellini/sort-es/commit/bdcdd9f240e5c0ae99b3024a42269b2f850493c4))
+* added sortable byStringLowerCase ([9410526](https://github.com/cosimochellini/sort-es/commit/9410526fe87ce8f6a136e084662f3d92456d97bf))
+
+
+### Bug Fixes
+
+* fiexd codecov CI ([e6749f0](https://github.com/cosimochellini/sort-es/commit/e6749f04278b90a519b2ef40d02337647eb8d71b))
+
 ### [0.2.11](https://github.com/cosimochellini/sort-es/compare/v0.2.10...v0.2.11) (2020-02-27)